byte text library integration

+++ cmd/eg/compute/compute.upload.go
@@ -161,7 +161,7 @@ func (t upload) Run(gctx *cmdopts.Global, tlsc *cmdopts.TLSConfig) (err error) {
 		Entry:       filepath.Join(ws.Module, filepath.Base(entry.Path)),
 		Ttl:         uint64(t.RuntimeResources.TTL.Milliseconds()),
 		Cores:       t.RuntimeResources.Cores,
-		Memory:      t.RuntimeResources.Memory,
+		Memory:      uint64(t.RuntimeResources.Memory),
 		Arch:        t.RuntimeResources.Arch,
 		Os:          t.RuntimeResources.OS,
 		AllowShared: t.HostedCompute,
